# src/main/resources/openapi/PrintAPIs.yaml
openapi: 3.0.0
info:
  title: Print APIs
  version: '1.24.0'
  description: Print APIs
#
# --------------------------------------------------------------------------------
#
paths:
...
  '/eros/{eroId}/temporary-certificates/applications/{applicationId}':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
      - name: applicationId
        description: The identifier of the Voter Card Application.
        schema:
          type: string
        in: path
        required: true
    options:
      deprecated: true
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - VAC Temporary Certificates
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    get:
      deprecated: true
      summary: Returns the Temporary Certificate summaries for an application. Deprecated. Use /eros/{eroId}/temporary-certificates?applicationId={applicationId} instead
      description: Returns the Temporary Certificate summaries for an application. Deprecated. Use /eros/{eroId}/temporary-certificates?applicationId={applicationId} instead
      tags:
        - VAC Temporary Certificates
      responses:
        200:
          $ref: '#/components/responses/TemporaryCertificateSummaries'
      operationId: deprecated-get-temporary-certificate-summaries-by-application-id
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/temporary-certificates/applications/{applicationId}'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.path.applicationId: method.request.path.applicationId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET

  #
  # Temporary Certificates
  # --------------------------------------------------------------------------------
  '/eros/{eroId}/temporary-certificates':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - VAC Temporary Certificates
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates and returns a temporary certificate
      description: Generates and returns a temporary certificate
      tags:
        - VAC Temporary Certificates
      requestBody:
        $ref: '#/components/requestBodies/GenerateTemporaryCertificate'
      responses:
        '201':
          $ref: '#/components/responses/PreSignedUrlResource'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: generate-temporary-certificate
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/temporary-certificates'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

    get:
      summary: Returns the Temporary Certificate summaries for a Voter Card Application
      description: Returns the Temporary Certificate summaries for a Voter Card Application
      parameters:
        - name: applicationId
          description: The identifier of the Voter Card Application.
          schema:
            type: string
          in: query
          required: true
      tags:
        - VAC Temporary Certificates
      responses:
        200:
          $ref: '#/components/responses/TemporaryCertificateSummaries'
      operationId: get-temporary-certificate-summaries-by-application-id
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/temporary-certificates'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: GET

  #
  # Temporary Certificates Explainer Document
  # --------------------------------------------------------------------------------
  '/eros/{eroId}/temporary-certificates/{gssCode}/explainer-document':
    parameters:
      - name: eroId
        description: The ID of the Electoral Registration Office responsible for the application.
        schema:
          type: string
        in: path
        required: true
      - name: gssCode
        description: The relevant GSS code for the local authority.
        schema:
          type: string
          example: 'E09000007'
        in: path
        required: true
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - VAC Temporary Certificates
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Returns the Local Authority's Explainer Document for a temporary certificate
      description: |
        Returns the Local Authority's Explainer Document for a temporary certificate  
        The Explainer Document is a separately generated and printed document that accompanies the Temporary Certificate
        and is used to explain to the Elector how to use their Temporary Certificate.
      tags:
        - VAC Temporary Certificates
      responses:
        '201':
          description: created
          headers:
            Content-Disposition:
              description: Content-Disposition header
              schema:
                type: string
                example: attachment; filename="temporary-certificate-explainer-document-E09000007.pdf"
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: generate-temporary-certificate-explainer-document
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/eros/{eroId}/temporary-certificates/{gssCode}/explainer-document'
        requestParameters:
          integration.request.path.eroId: method.request.path.eroId
          integration.request.path.gssCode: method.request.path.gssCode
          integration.request.header.x-request-id: context.requestId
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST


  #
  # Anonymous Elector Document Summaries
  # --------------------------------------------------------------------------------
...
components:
...
  schemas:
    ErrorResponse:
      title: ErrorResponse
      description: |
        A generic response body object describing errors in a web request, and can be used to communicate several
        different types of error condition such as (but not limited to) `400 BAD REQUEST`, `409 CONFLICT` etc.
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2022-09-28T18:01:42.105Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'Validation failed for object=applicationRequest. Error count: 14'
        validationErrors:
          description: Validation errors are only present if the error being described is a `400 BAD REQUEST`.
          type: array
          items:
            type: string
          example: 'Error on field applicant.nino: rejected value [aaaaaaaaaaa], must match ^.{1,10}$'
      required:
        - timestamp
        - status
        - error
        - message
...
    SourceType:
      title: SourceType
      description: An enumeration of sources
      type: string
      enum:
        - voter-card
        - anonymous-elector-document
    CertificateLanguage:
      title: CertificateLanguage
      description: An enumeration of Certificate languages
      type: string
      enum:
        - cy
        - en
      default: en
...
    TemporaryCertificateSummariesResponse:
      title: TemporaryCertificateSummariesResponse
      type: object
      description: A response containing a collection of Temporary Certificate summaries
      properties:
        temporaryCertificates:
          type: array
          description: List of Temporary Certificate summaries for the application.
          items:
            $ref: '#/components/schemas/TemporaryCertificateSummary'
      required:
        - temporaryCertificates
    TemporaryCertificateSummary:
      title: TemporaryCertificateSummary
      type: object
      description: An object containing Temporary Certificate summary
      properties:
        certificateNumber:
          type: string
          pattern: '^[0-9AC-HJ-NP-RT-Z]{20}$'
          description: The unique identifier of the Certificate. A 20 character alphanumeric string.
          example: DEV1W0XDH1368LWE1J40
        status:
          $ref: '#/components/schemas/TemporaryCertificateStatus'
        userId:
          type: string
          format: email
          example: fred.blogs@some-domain.co.uk
          description: UserId of the user who generated the Temporary Certificate
        dateTimeGenerated:
          type: string
          format: date-time
          description: The timestamp in ISO-8601 UTC that the Temporary Certificate was generated
          example: '2022-11-17T10:53:25Z'
        issueDate:
          type: string
          format: date
          description: The date that the Temporary Certificate was issued
          example: '2022-11-17'
        validOnDate:
          type: string
          format: date
          description: The date on which the Temporary Certificate is/was valid
          example: '2022-11-19'
      required:
        - certificateNumber
        - status
        - userId
        - dateTimeGenerated
        - issueDate
        - validOnDate
    TemporaryCertificateStatus:
      title: TemporaryCertificateStatus
      type: string
      enum:
        - generated
    GenerateTemporaryCertificateRequest:
      title: GenerateTemporaryCertificateRequest
      description: Request body to generate a Temporary Certificate
      type: object
      properties:
        gssCode:
          type: string
          example: E12345678
          minLength: 9
          maxLength: 9
          description: GSS code of the ERO responsible for the application
        sourceType:
          $ref: '#/components/schemas/SourceType'
        sourceReference:
          type: string
          example: 1f0f76a9a66f438b9bb33070
          description: Reference in the source application that this Temporary Certificate is for.
        applicationReference:
          type: string
          example: V3JSZC4CRH
          description: The application reference as known by the citizen. Not guaranteed to be unique.
        firstName:
          type: string
          maxLength: 255
          description: First name of the Elector
          example: John
        middleNames:
          type: string
          maxLength: 255
          description: Middle names of the Elector
          example: Malcolm
        surname:
          type: string
          maxLength: 255
          description: Surname of the Elector
          example: Smith
        certificateLanguage:
          $ref: '#/components/schemas/CertificateLanguage'
        photoLocation:
          type: string
          description: |
            The location of the Elector's Certificate photo. 
            This is the photo that has been prepared and approved for the Certificate by the vc-admin, rather than the photo that the applicant uploaded as part of their original application. 
            Typically an S3 ARN
          maxLength: 1024
        validOnDate:
          type: string
          format: date
          description: The date that the Temporary Certificate is valid for.
          example: '2023-07-27'
        allowLargeResponse:
          type: boolean
      required:
        - gssCode
        - sourceType
        - sourceReference
        - applicationReference
        - firstName
        - surname
        - certificateLanguage
        - photoLocation
        - validOnDate
...
    PreSignedUrlResourceResponse:
      title: PreSignedUrlResourceResponse
      description: An object detailing the pre-signed URL to use in order to retrieve the resource
      properties:
        preSignedUrl:
          type: string
          format: uri
          description: AWS pre-signed URL giving time-limited access to the resource.
          example: 'https://example.s3.eu-west-2.amazonaws.com/image.png?X-Amz-Expires=1800'
      required:
        - preSignedUrl

  #
  # Response Body Definitions
  # --------------------------------------------------------------------------------
  responses:
    400ErrorResponse:
      description: Error response for a HTTP 400 describing fields in the request payload that are in error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2022-09-28T18:01:42.105Z'
            status: 400
            error: 'Bad Request'
            message: 'Validation failed for object=applicationRequest. Error count: 14'
            validationErrors: 'Error on field applicant.nino: rejected value [aaaaaaaaaaa], must match ^.{1,10}$'
    404ErrorResponse:
      description: |
        Error response for a HTTP 404 describing the problem with the request.
        Typical examples are resources such as the application not found.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2022-09-28T18:01:42.105Z'
            status: 404
            error: 'Not Found'
            message: 'Certificate photo for voter card application with eroId=[some-ero] and applicationType=[VOTER_CARD] and applicationId=[507f1f77bcf86cd799439011] not found'
...
    TemporaryCertificateSummaries:
      description: Response containing a collection of Temporary Certificate summaries
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TemporaryCertificateSummariesResponse'
...
    PreSignedUrlResource:
      description: Response containing an object detailing the pre-signed URL for the resource
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PreSignedUrlResourceResponse'

  #
  # Request Body Definitions
  # --------------------------------------------------------------------------------
  requestBodies:
    GenerateTemporaryCertificate:
      description: Request body to generate a Temporary Certificate
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateTemporaryCertificateRequest'
...
  securitySchemes:
    eroUserCognitoUserPoolAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        providerARNs:
          - '${cognito_ero_user_pool_arn}'
        type: cognito_user_pools
    awsIamAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4